KAFKA-17626: Move common fetch related classes in storage to storage API
