+++ build.gradle
@@ -954,7 +954,7 @@ project(':share') {
 
   dependencies {
     implementation project(':server-common')
-    implementation project(':storage')
+    implementation project(':storage:storage-api')
 
     implementation libs.slf4jApi
