+++ checkstyle/import-control-share.xml
@@ -45,7 +45,7 @@
       <allow pkg="org.apache.kafka.server.share" />
 
       <subpackage name="fetch">
-        <allow class="org.apache.kafka.storage.internals.log.FetchParams"/>
+        <allow class="org.apache.kafka.server.log.storage.FetchParams"/>
       </subpackage>
     </subpackage>
   </subpackage>
