+++ core/src/main/scala/kafka/coordinator/group/CoordinatorLoaderImpl.scala
@@ -25,8 +25,8 @@ import org.apache.kafka.common.requests.TransactionResult
 import org.apache.kafka.common.utils.Time
 import org.apache.kafka.coordinator.common.runtime.CoordinatorLoader.{LoadSummary, UnknownRecordTypeException}
 import org.apache.kafka.coordinator.common.runtime.{CoordinatorLoader, CoordinatorPlayback, Deserializer}
+import org.apache.kafka.server.log.storage.FetchIsolation
 import org.apache.kafka.server.util.KafkaScheduler
-import org.apache.kafka.storage.internals.log.FetchIsolation
 
 import java.nio.ByteBuffer
 import java.util.concurrent.CompletableFuture
