+++ core/src/main/scala/kafka/server/DelayedFetch.scala
@@ -25,8 +25,9 @@ import org.apache.kafka.common.errors._
 import org.apache.kafka.common.protocol.Errors
 import org.apache.kafka.common.requests.FetchRequest.PartitionData
 import org.apache.kafka.common.requests.OffsetsForLeaderEpochResponse.{UNDEFINED_EPOCH, UNDEFINED_EPOCH_OFFSET}
+import org.apache.kafka.server.log.storage.{FetchIsolation, FetchParams, FetchPartitionData}
 import org.apache.kafka.server.metrics.KafkaMetricsGroup
-import org.apache.kafka.storage.internals.log.{FetchIsolation, FetchParams, FetchPartitionData, LogOffsetMetadata}
+import org.apache.kafka.storage.internals.log.LogOffsetMetadata
 
 import scala.collection._
 import scala.jdk.CollectionConverters._
