+++ core/src/test/java/kafka/server/share/SharePartitionManagerTest.java
@@ -49,6 +49,8 @@ import org.apache.kafka.common.requests.ShareRequestMetadata;
 import org.apache.kafka.common.utils.ImplicitLinkedHashCollection;
 import org.apache.kafka.common.utils.MockTime;
 import org.apache.kafka.common.utils.Time;
+import org.apache.kafka.server.log.storage.FetchIsolation;
+import org.apache.kafka.server.log.storage.FetchParams;
 import org.apache.kafka.server.share.CachedSharePartition;
 import org.apache.kafka.server.share.ErroneousAndValidPartitionData;
 import org.apache.kafka.server.share.NoOpShareStatePersister;
@@ -68,8 +70,6 @@ import org.apache.kafka.server.util.timer.SystemTimer;
 import org.apache.kafka.server.util.timer.SystemTimerReaper;
 import org.apache.kafka.server.util.timer.Timer;
 import org.apache.kafka.storage.internals.log.FetchDataInfo;
-import org.apache.kafka.storage.internals.log.FetchIsolation;
-import org.apache.kafka.storage.internals.log.FetchParams;
 import org.apache.kafka.storage.internals.log.LogOffsetMetadata;
 import org.apache.kafka.test.TestUtils;
