+++ core/src/test/java/kafka/server/share/SharePartitionTest.java
@@ -37,6 +37,7 @@ import org.apache.kafka.common.record.MemoryRecordsBuilder;
 import org.apache.kafka.common.record.TimestampType;
 import org.apache.kafka.common.utils.MockTime;
 import org.apache.kafka.common.utils.Time;
+import org.apache.kafka.server.log.storage.FetchPartitionData;
 import org.apache.kafka.server.share.NoOpShareStatePersister;
 import org.apache.kafka.server.share.PartitionFactory;
 import org.apache.kafka.server.share.Persister;
@@ -49,7 +50,6 @@ import org.apache.kafka.server.util.FutureUtils;
 import org.apache.kafka.server.util.timer.SystemTimer;
 import org.apache.kafka.server.util.timer.SystemTimerReaper;
 import org.apache.kafka.server.util.timer.Timer;
-import org.apache.kafka.storage.internals.log.FetchPartitionData;
 import org.apache.kafka.test.TestUtils;
 
 import org.junit.jupiter.api.AfterEach;
