+++ core/src/test/scala/integration/kafka/server/DelayedRemoteFetchTest.scala
@@ -22,6 +22,7 @@ import org.apache.kafka.common.protocol.{ApiKeys, Errors}
 import org.apache.kafka.common.record.MemoryRecords
 import org.apache.kafka.common.requests.FetchRequest
 import org.apache.kafka.common.{TopicIdPartition, Uuid}
+import org.apache.kafka.server.log.storage.{FetchIsolation, FetchParams, FetchPartitionData}
 import org.apache.kafka.server.metrics.KafkaYammerMetrics
 import org.apache.kafka.storage.internals.log._
 import org.junit.jupiter.api.Assertions._
