+++ storage/api/src/main/java/org/apache/kafka/server/log/storage/FetchParams.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.storage.internals.log;
+package org.apache.kafka.server.log.storage;
 
 import org.apache.kafka.common.replica.ClientMetadata;
 import org.apache.kafka.common.requests.FetchRequest;
