+++ storage/src/main/java/org/apache/kafka/storage/internals/log/RemoteStorageFetchInfo.java
@@ -18,6 +18,7 @@ package org.apache.kafka.storage.internals.log;
 
 import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.requests.FetchRequest;
+import org.apache.kafka.server.log.storage.FetchIsolation;
 
 public class RemoteStorageFetchInfo {
